# workstation-rack/src/main/resources/META-INF/openapi.yaml
...
openapi: 3.0.3
info:
  title: workstation-rack API
  version: 1.0.0-SNAPSHOT
paths:
...
  /workstation/booking/{id}:
    get:
      tags:
      - Booking Resource
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "200":
          description: OK
    put:
      tags:
      - Booking Resource
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Booking"
      responses:
        "200":
          description: OK
    delete:
      tags:
      - Booking Resource
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "200":
          description: OK
...
  /workstation/racks/{id}:
    get:
      tags:
      - Rack Resource
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "200":
          description: OK
    put:
      tags:
      - Rack Resource
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Rack"
      responses:
        "200":
          description: OK
    delete:
      tags:
      - Rack Resource
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "200":
          description: OK
...
components:
  schemas:
    Booking:
      type: object
      properties:
        id:
          format: int64
          type: integer
        rackId:
          format: int64
          type: integer
        requesterId:
          format: int64
          type: integer
        bookFrom:
          $ref: "#/components/schemas/LocalDateTime"
        bookTo:
          $ref: "#/components/schemas/LocalDateTime"
...
    LocalDateTime:
      format: date-time
      type: string
      example: 2022-03-10T12:15:50
    Rack:
      type: object
      properties:
        id:
          format: int64
          type: integer
        serialNumber:
          type: string
        status:
          type: string
        teamId:
          format: int64
          type: integer
...